Refactor command and execution handling; introduce execution context, update argument structures, and enhance error handling in dispatch logic.

+++ include/parser_state.h
@@ -14,5 +14,5 @@ typedef struct parser_state{
     parser_result result;
 }parser_state;
 
-void parser_state_init(parser_state* parser_state ,char* b_ptr, u32 end_idx);
+void parser_state_init(parser_state* parser_state ,char* b_ptr, size_t end_idx);
 void parser_state_reset(parser_state* parser_state);
